requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.2
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pydantic[email]==2.10.2
python-dotenv==1.0.1
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
bcrypt==4.2.1
uvicorn[standard]==0.30.3
email-validator==2.2.0
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
